:art: move withLoader from nt-core to nt-utils
